+++ Sheet/index.js
@@ -13,9 +13,6 @@ import Overlay from '../Overlay';
 const NOOP = () => {};
 
 const styles = StyleSheet.create({
-  modal: {
-    position: 'relative',
-  },
   container: {
     position: 'absolute',
     bottom: 0,
@@ -94,8 +91,8 @@ class Sheet extends Component {
     return (
       <Overlay
         visible={this.state.visible}
-        style={[styles.modal, this.props.modalStyle]}
-        onPress={this.props.onPressModal}
+        style={this.props.overlayStyle}
+        onPress={this.props.onPressOverlay}
       >
         <View style={styles.container}>
           <Animated.View
@@ -118,11 +115,11 @@ Sheet.propTypes = {
   // 显示开关
   visible: PropTypes.bool.isRequired,
   // 遮罩层样式
-  modalStyle: View.propTypes.style,
+  overlayStyle: View.propTypes.style,
   // 关闭回调（动画结束时）
   onClose: PropTypes.func,
   // 遮罩点击事件
-  onPressModal: PropTypes.func,
+  onPressOverlay: PropTypes.func,
   // 子元素
   children: PropTypes.oneOfType([PropTypes.element, PropTypes.array]),
   // 动画时长
@@ -131,9 +128,9 @@ Sheet.propTypes = {
 Sheet.defaultProps = {
   style: null,
   visible: false,
-  modalStyle: null,
+  overlayStyle: null,
   onClose: NOOP,
-  onPressModal: NOOP,
+  onPressOverlay: NOOP,
   children: null,
   duration: 200,
 };